+++ src/scripts/attach.ts
@@ -141,12 +141,32 @@ async function getCurrentDebugPort() {
 }
 
 
+// TODO: Remove this function when the deprecated setting is removed
+function getDeprecatedSetting<T>(key: string, deprecatedKey: string): T | undefined {
+    const extConfig = utils.getExtensionConfig();
+
+    const value = extConfig.get<T>(key);
+    if (value !== undefined && value !== extConfig.inspect(key)?.defaultValue) {
+        return value;
+    }
+
+    const deprecatedValue = extConfig.get<T>(deprecatedKey);
+    if (deprecatedValue !== undefined && deprecatedValue !== extConfig.inspect(deprecatedKey)?.defaultValue) {
+        vscode.window.showWarningMessage(`The setting 'motionbuilder.${deprecatedKey}' is deprecated, please use 'motionbuilder.${key}' instead`);
+        return deprecatedValue;
+    }
+
+    return value;
+}
+
+
 /**
  * Get a free port to use for the debugpy server
  */
 async function getWantedPort() {
     const extConfig = utils.getExtensionConfig();
-    const port = extConfig.get<number>("debug.port");
+    // const port = extConfig.get<number>("attach.port");
+    const port = getDeprecatedSetting<number>("attach.port", "debug.port");
     if (!port) {
         return null;
     }
@@ -155,7 +175,8 @@ async function getWantedPort() {
         return port;
     }
 
-    if (extConfig.get<boolean>("debug.autoPort")) {
+    // if (extConfig.get<boolean>("attach.autoPort")) {
+    if (getDeprecatedSetting<boolean>("attach.autoPort", "debug.autoPort")) {
         const freePort = await utils.findFreePort(port, 100);
         if (freePort) {
             return freePort;